requirement.txt
pyodbc
mssql-django